Lab 5 WORKING!

+++ engine/render/MeshResource.h
@@ -13,7 +13,7 @@ class MeshResource
 public:
 	static std::shared_ptr<MeshResource> Cube();
 	static std::shared_ptr<MeshResource> LoadObj(const char* pathToFile);
-	MeshResource(Vertex vertices[], uint64_t verticeslength, uint64_t indices[], uint64_t indicesLength);
+	MeshResource(Vertex vertices[], uint32_t verticeslength, uint32_t indices[], uint32_t indicesLength);
 	~MeshResource();
 	void Destroy();
 	void render();
